// src/app/specialty/interfaces/specialty.interface.ts
// Generated by https://quicktype.io

export interface SearchSpecialty {
  data:  Specialty[];
  meta:  Meta;
  type:  string;
  links: Links;
}

export interface Specialty {
  id:         number;
  name:       string;
  status:     number;
  created_at: string;
}

export interface Links {
  first: string;
  last:  string;
  prev:  null;
  next:  null;
}

export interface Meta {
  organization: string;
  authors:      string[];
  current_page: number;
  from:         number;
  last_page:    number;
  links:        Link[];
  path:         string;
  per_page:     number;
  to:           number;
  total:        number;
}

export interface Link {
  url:    null | string;
  label:  string;
  active: boolean;
}