// src/app/zone/interfaces/zone.interface.ts
// Generated by https://quicktype.io

export interface SearchZone {
  data:  Zone[];
  meta:  Meta;
  type:  string;
  links: Links;
}

export interface Zone {
  id:         number;
  name:       string;
  status:     number;
  created_at: string;
}

export interface Links {
  first: string;
  last:  string;
  prev:  null;
  next:  null;
}

export interface Meta {
  organization: string;
  authors:      string[];
  current_page: number;
  from:         number;
  last_page:    number;
  links:        Link[];
  path:         string;
  per_page:     number;
  to:           number;
  total:        number;
}

export interface Link {
  url:    null | string;
  label:  string;
  active: boolean;
}