// src/app/teacher/interfaces/teacher.interface.ts
// Generated by https://quicktype.io

export interface SearchTeacher {
  data:  Teacher[];
  meta:  Meta;
  type:  string;
  links: Links;
}

export interface Teacher {
  id:         number;
  names:      string;
  lastNames:  string;
  address:    string;
  dui:        string;
  nip:        string;
  nit:        string;
  nup:        string;
  career:     string;
  category:   string;
  position:   string;
  departure:  string;
  subLevel:   string;
  entryDate:  string;
  status:     string;
  created_at: string;
}

export interface Links {
  first: string;
  last:  string;
  prev:  null;
  next:  null;
}

export interface Meta {
  organization: string;
  authors:      string[];
  current_page: number;
  from:         number;
  last_page:    number;
  links:        Link[];
  path:         string;
  per_page:     number;
  to:           number;
  total:        number;
}

export interface Link {
  url:    null | string;
  label:  string;
  active: boolean;
}